Fix cmd-duration flag overflow error
- Change cmdDuration from uint16 to uint32 to handle longer command durations
- Update DurationConfig.MinTime from uint16 to uint32 for consistency
- Update duration renderer to handle uint32 type
- Fixes 'value out of range' error for durations > 65535ms

Fixes issue where commands taking longer than ~65 seconds would cause
a parsing error due to uint16 overflow.

+++ config/config.go
@@ -37,7 +37,7 @@ type CommandNumberConfig struct {
 // DurationConfig ...
 type DurationConfig struct {
 	BaseComponentConfig
-	MinTime uint16 `toml:"min_time"`
+	MinTime uint32 `toml:"min_time"`
 }
 
 // DirectoryConfig ...
